Combine DQA files and issues into one tool
To support new features for each tool, there is an overlap of functionality.
To simplify further development and usage, the code has been combined.

The pedsnet-dqa binary will be generated that supports subcommands. The
files program is now called generate and the issues program is now called
report.
